+++ src/platforms/browser/WebPlatform.mjs
@@ -90,7 +90,7 @@ export default class WebPlatform {
 
     loop() {
         let self = this;
-        let lp = function() {
+        let lp = function () {
             self._awaitingLoop = false;
             if (self._looping) {
                 self.stage.updateFrame();
@@ -103,6 +103,23 @@ export default class WebPlatform {
             }
         }
         requestAnimationFrame(lp);
+    }
+
+    uploadCompressedGlTexture(gl, textureSource, source, options) {
+        const view = !source.pvr ? new DataView(source.mipmaps[0]) : source.mipmaps[0];
+        gl.compressedTexImage2D(
+            gl.TEXTURE_2D,
+            0,
+            source.glInternalFormat,
+            source.pixelWidth,
+            source.pixelHeight,
+            0,
+            view,
+        )
+        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE);
+        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE);
+        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.LINEAR);
+        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.LINEAR);
     }
 
     uploadGlTexture(gl, textureSource, source, options) {
@@ -124,24 +141,155 @@ export default class WebPlatform {
         }
     }
 
-    loadSrcTexture({src, hasAlpha}, cb) {
+    /**
+     * KTX File format specification
+     * https://www.khronos.org/registry/KTX/specs/1.0/ktxspec_v1.html
+     **/
+    handleKtxLoad(cb, src) {
+        var self = this;
+        return function () {
+            var arraybuffer = this.response;
+            var view = new DataView(arraybuffer);
+
+            // identifier, big endian
+            var targetIdentifier = 3632701469
+            if (targetIdentifier !== (view.getUint32(0) + view.getUint32(4) + view.getUint32(8))) {
+                cb('Parsing failed: identifier ktx mismatch:', src)
+            }
+
+            var littleEndian = (view.getUint32(12) === 16909060) ? true : false;
+            var data = {
+                glType: view.getUint32(16, littleEndian),
+                glTypeSize: view.getUint32(20, littleEndian),
+                glFormat: view.getUint32(24, littleEndian),
+                glInternalFormat: view.getUint32(28, littleEndian),
+                glBaseInternalFormat: view.getUint32(32, littleEndian),
+                pixelWidth: view.getUint32(36, littleEndian),
+                pixelHeight: view.getUint32(40, littleEndian),
+                pixelDepth: view.getUint32(44, littleEndian),
+                numberOfArrayElements: view.getUint32(48, littleEndian),
+                numberOfFaces: view.getUint32(52, littleEndian),
+                numberOfMipmapLevels: view.getUint32(56, littleEndian),
+                bytesOfKeyValueData: view.getUint32(60, littleEndian),
+                kvps: [],
+                mipmaps: [],
+                get width() { return this.pixelWidth },
+                get height() { return this.pixelHeight },
+            };
+
+            const props = (obj) => {
+                const p = [];
+                for (let v in obj) {
+                    p.push(obj[v]);
+                }
+                return p;
+            }
+
+            const formats = Object.values(self.stage.renderer.getCompressedTextureExtensions())
+                .filter((obj) => obj != null)
+                .map((obj) => props(obj))
+                .reduce((prev, current) => prev.concat(current));
+
+            if (!formats.includes(data.glInternalFormat)) {
+                console.warn("[Lightning] Unrecognized texture extension format:", src, data.glInternalFormat, self.stage.renderer.getCompressedTextureExtensions());
+            }
+
+            var offset = 64
+            // Key Value Pairs of data start at byte offset 64
+            // But the only known kvp is the API version, so skipping parsing.
+            offset += data.bytesOfKeyValueData;
+
+            for (var i = 0; i < data.numberOfMipmapLevels; i++) {
+                var imageSize = view.getUint32(offset);
+                offset += 4;
+                data.mipmaps.push(view.buffer.slice(offset, imageSize));
+                offset += imageSize
+            }
+
+            cb(null, {
+                source: data,
+                renderInfo: { src: src, compressed: true },
+            })
+        }
+    }
+
+    handlePvrLoad(cb, src) {
+        return function () {
+            // pvr header length in 32 bits
+            const pvrHeaderLength = 13;
+            // for now only we only support: COMPRESSED_RGB_ETC1_WEBGL
+            const pvrFormatEtc1 = 0x8D64;
+            const pvrWidth = 7;
+            const pvrHeight = 6;
+            const pvrMipmapCount = 11;
+            const pvrMetadata = 12;
+            const arrayBuffer = this.response;
+            const header = new Int32Array(arrayBuffer, 0, pvrHeaderLength);
+            const dataOffset = header[pvrMetadata] + 52;
+            const pvrtcData = new Uint8Array(arrayBuffer, dataOffset);
+
+            var data = {
+                glInternalFormat: pvrFormatEtc1,
+                pixelWidth: header[pvrWidth],
+                pixelHeight: header[pvrHeight],
+                numberOfMipmapLevels: header[pvrMipmapCount],
+                mipmaps: [],
+                pvr: true,
+                get width() { return this.pixelWidth },
+                get height() { return this.pixelHeight },
+            };
+
+            let offset = 0
+            let width = data.pixelWidth;
+            let height = data.pixelHeight;
+
+            for (var i = 0; i < data.numberOfMipmapLevels; i++) {
+                const level = ((width + 3) >> 2) * ((height + 3) >> 2) * 8;
+                const view = new Uint8Array(arrayBuffer, pvrtcData.byteOffset + offset, level);
+                data.mipmaps.push(view);
+                offset += level;
+                width = width >> 1;
+                height = height >> 1;
+            }
+
+            cb(null, {
+                source: data,
+                renderInfo: { src: src, compressed: true },
+            })
+        }
+    }
+
+    loadSrcTexture({ src, hasAlpha }, cb) {
         let cancelCb = undefined;
         let isPng = (src.indexOf(".png") >= 0) || src.substr(0, 21) == 'data:image/png;base64';
-        if (this._imageWorker) {
+        let isKtx = src.indexOf('.ktx') >= 0;
+        let isPvr = src.indexOf('.pvr') >= 0;
+        if (isKtx || isPvr) {
+            let request = new XMLHttpRequest();
+            request.addEventListener(
+                "load", isKtx ? this.handleKtxLoad(cb, src) : this.handlePvrLoad(cb, src)
+            );
+            request.open("GET", src);
+            request.responseType = "arraybuffer";
+            request.send();
+            cancelCb = function () {
+                request.abort();
+            }
+        } else if (this._imageWorker) {
             // WPE-specific image parser.
             const image = this._imageWorker.create(src);
-            image.onError = function(err) {
+            image.onError = function (err) {
                 return cb("Image load error");
             };
-            image.onLoad = function({imageBitmap, hasAlphaChannel}) {
+            image.onLoad = function ({ imageBitmap, hasAlphaChannel }) {
                 cb(null, {
                     source: imageBitmap,
-                    renderInfo: {src: src},
+                    renderInfo: { src: src, compressed: false },
                     hasAlpha: hasAlphaChannel,
                     premultiplyAlpha: true
                 });
             };
-            cancelCb = function() {
+            cancelCb = function () {
                 image.cancel();
             }
         } else {
@@ -149,26 +297,26 @@ export default class WebPlatform {
 
             // On the PS4 platform setting the `crossOrigin` attribute on
             // images can cause CORS failures.
-            if (!(src.substr(0,5) == "data:") && !Utils.isPS4) {
+            if (!(src.substr(0, 5) == "data:") && !Utils.isPS4) {
                 // Base64.
                 image.crossOrigin = "Anonymous";
             }
-            image.onerror = function(err) {
+            image.onerror = function (err) {
                 // Ignore error message when cancelled.
                 if (image.src) {
                     return cb("Image load error");
                 }
             };
-            image.onload = function() {
+            image.onload = function () {
                 cb(null, {
                     source: image,
-                    renderInfo: {src: src},
+                    renderInfo: { src: src, compressed: false },
                     hasAlpha: isPng || hasAlpha
                 });
             };
             image.src = src;
 
-            cancelCb = function() {
+            cancelCb = function () {
                 image.onerror = null;
                 image.onload = null;
                 image.removeAttribute('src');