+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v2.8.0
+
+*20 oct 2022*
+
+- Updated patching.md file (#422)
+- Minor documentation updates (#405)
+- Added support to track vram usage (#395)
+- Added texture compression support (#391)
+
 ## v2.7.1
 
 *19 sep 2022*
@@ -332,4 +341,5 @@
 
 ### Fixes
 * Fixed text rendering artifacts sometimes appearing on RPI platform (#41)
+